fix: setup関数をhandlerにフォルダに移動する
